Day 2: Added in Commands, Statements, Errors, Factories, Processors
Added Command and Statement enum types as well as processors for different command types.

Will add builders for "building" statements.

Data Flow:
Tokens -> Commands -> Statements
CommandProcessors convert Tokens into Commands with a CommandType as well as the specific type of statement it will be with a "marker".

FactoryRouter then routes the command to it's correlating factory to create a statement using the marker.

The statements will then be built using builders. (This will be done by having each statement hold a query of relevant data as all statements may be different, will utilize polymorphism).